+++ src/nodes/effect/location.ts
@@ -5,7 +5,7 @@ import { ALIGN_OPTIONS, EDGE_OPTIONS } from "./constants";
 
 type LocationKind = {
 	points: Point;
-	targets: any;
+	targets: PositionSource;
 };
 
 type TInputs = {
@@ -58,7 +58,7 @@ class LocationNode extends EffectModifierNode<TInputs, TState> {
 		return [
 			this.effectInput,
 			{ key: "location", type: "point", ...this.io("location"), state: "points" },
-			{ key: "location", type: "any", ...this.io("location"), state: "targets" },
+			{ key: "location", type: "position", ...this.io("location"), state: "targets" },
 			{ key: "attachTo", type: "boolean", ...this.sharedIo("attachTo"), state: "targets" },
 			{ key: "cacheLocation", type: "boolean", ...this.sharedIo("cacheLocation"), state: "targets" },
 			{ key: "cacheLocation", type: "boolean", ...this.sharedIo("cacheLocation"), state: "points" },